Updated Navigation - Removed the Nagivation components that became depricated - Moved the new main Navigation.tsx into the navigation directory
