--- src/navigation/HomeNavigation.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { createStackNavigator } from "@react-navigation/stack";
-import HomeScreen from "../screens/HomeScreen";
-
-const Stack = createStackNavigator();
-
-const HomeNavigation = () => {
-	return (
-		<Stack.Navigator>
-			<Stack.Screen name="Home" component={HomeScreen} />
-		</Stack.Navigator>
-	);
-};
-
-export default HomeNavigation;
